little update still ULONGMAX

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: olamrabt <olamrabt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/01 18:33:49 by olamrabt          #+#    #+#             */
-/*   Updated: 2023/12/03 14:03:54 by olamrabt         ###   ########.fr       */
+/*   Updated: 2023/12/03 19:28:09 by olamrabt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,13 +17,14 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <stdio.h>
+#include <limits.h>
 
 // main func
 int ft_printf(const char *, ...);
 // helper functions
 int handle_spe(char sp, va_list ap);
 int print_str(char *str);
-int print_nbr_base(long nbr, int base, int to_upper);
+int print_nbr_base(long long nbr, int base, int to_upper);
 // • %c Prints a single character.
 // • %s Prints a string (as defined by the common C convention).
 // • %p The void * pointer argument has to be printed in hexadecimal format.
